moving the command to release management
