+++ BREEZE.rst
@@ -466,11 +466,11 @@ Publishing the documentation
 --------------------------
 
 To publish the documentation generated by ``build-docs`` in Breeze to ``airflow-site``,
-use the ``publish-docs`` command:
+use the ``release-management publish-docs`` command:
 
 .. code-block:: bash
 
-     breeze publish-docs
+     breeze release-management publish-docs
 
 The publishing documentation consists  steps:
 
@@ -480,7 +480,7 @@ The publishing documentation consists  steps:
 
 .. code-block:: bash
 
-     breeze publish-docs --package-filter apache-airflow-providers-amazon
+     breeze release-management publish-docs --package-filter apache-airflow-providers-amazon
 
 The flag ``--package-filter`` can be used to selectively publish docs during a release. It can take
 values such as apache-airflow, helm-chart, apache-airflow-providers, or any individual providers.
@@ -488,22 +488,22 @@ The documentation publication happens based on this flag.
 
 .. code-block:: bash
 
-     breeze publish-docs --override-versioned
+     breeze release-management publish-docs --override-versioned
 
 The flag ``--override-versioned`` is a boolean flag that is used to override the versioned directories
 while publishing the documentation.
 
 .. code-block:: bash
 
-     breeze publish-docs --airflow-site-directory
+     breeze release-management publish-docs --airflow-site-directory
 
 The flag ``--airflow-site-directory`` takes the path of the cloned ``airflow-site``. The command will
 not proceed if this is an invalid path.
 
-Those are all available flags of ``publish-docs`` command:
+Those are all available flags of ``release-management publish-docs`` command:
 
-.. image:: ./images/breeze/output_publish-docs.svg
-  :target: https://raw.githubusercontent.com/apache/airflow/main/images/breeze/output_publish-docs.svg
+.. image:: ./images/breeze/output_release-management_publish-docs.svg
+  :target: https://raw.githubusercontent.com/apache/airflow/main/images/breeze/output_release-management_publish-docs.svg
   :width: 100%
   :alt: Breeze Publish documentation
