+++ images/breeze/output-commands-hash.txt
@@ -43,22 +43,22 @@ release-management:install-provider-packages:a89493be1ae961c13469b5a25a605069
 release-management:prepare-airflow-package:85d01c57e5b5ee0fb9e5f9d9706ed3b5
 release-management:prepare-provider-documentation:bc0b95c6a0f308273ab25f728afdb039
 release-management:prepare-provider-packages:3e19a787c71d5515a2bc05539da0022f
+release-management:publish-docs:4a7d78a846f7e58c6af4a55cbdf7104b
 release-management:release-prod-images:4d85a23a2175bf8894de5aedbdd85614
 release-management:start-rc-process:108c9daabd5970fde86210b1669990cc
 release-management:start-release:ad13846c7573ddaaa59adc72f0f00284
 release-management:verify-provider-packages:96dce5644aad6b37080acf77b3d8de3a
-release-management:11bb87706beda348eec0182fb9138eb4
+release-management:473d554e19e108635deda89eb8e0bd05
 sbom:generate-provider-requirements:1e7609a0ebafb839d30ec90e4254a628
 sbom:update-sbom-information:0ce56884e5f842e3e80d6619df1ccc64
 sbom:d9c57d78e23a29ec9cd0139e230e1811
 setup:autocomplete:fffcd49e102e09ccd69b3841a9e3ea8e
-setup:check-all-params-in-groups:fa96f067af2bd80528ffd3f872ba6170
+setup:check-all-params-in-groups:c2ee7aa0ac0831c92167ed70c05f3bd3
 setup:config:38ebaaf93ed42bc7b2a3000eeea2631d
-setup:regenerate-command-images:a688d2d0ebb7e8cd88f94804257973f4
+setup:regenerate-command-images:76cc7b6568a1488a25e54fd5b4506b0a
 setup:self-upgrade:4af905a147fcd6670a0e33d3d369a94b
 setup:version:be116d90a21c2afe01087f7609774e1e
-setup:201c30ea237ea013a9a209a77092a2e8
-setup:201c30ea237ea013a9a209a77092a2e8
+setup:d5f6d65e2da4274ee76abf712e7f699e
 shell:48db0babd831d704b97118d053ee639a
 start-airflow:52ad9113f74b39c1ffbd37464e1415e6
 static-checks:6ba0614c2c493b45bfc027c3570d737e
